requirements.txt
Flask
line-bot-sdk
openai
werkzeug
langchain
langchain_community
langchain_openai
faiss-cpu
pypdf